requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-multipart
openai
tenacity
httpx
structlog
redis
pypdf
python-docx
beautifulsoup4
lxml
rank-bm25
faiss-cpu
sentence-transformers
langgraph
langsmith
ragas
datasets
pandas
numpy
streamlit
pytest
pytest-asyncio